requirements.txt
Flask-Migrate==3.1.0
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
openai==1.43.0
flask_cors==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.22